pyproject.toml
[project]
name = "spatial"
version = "0.1.0"
description = ""
authors = [
    {name = "Jacob Rogers",email = "jacob.p.r@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce (>=2.5.3,<3.0.0)",
    "ruff (>=0.11.6,<0.12.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "structlog (>=25.2.0,<26.0.0)",
    "structlog-round (>=1.0,<2.0)",
]

[tool.poetry]
packages = [
    { include = "spatial" },
    { include = "pygame_me" },
]

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pytest = "^8.3.5"
types-pyyaml = "^6.0.12.20250402"
rich = "^14.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
extend-select = [
  "ALL"
  # "F", # pyflakes
  # "E", # pycodestyle
  # "W", # pycodestyle warnings
  # "I", # isort
  # "N", # pep8-naming
  # "B", # flake8-bugbear
]
ignore = [
  # pycodestyle (for black)
  "E203",
  # pycodestyle (overlaps with bugbear)
  "E722",
  # Allow ### comments
  "E266",
  # Do not require blank line before class docstring
  "D203",
  # Docstring starts on first line.
  "D213",
  "D204",
  # Allow Assert
  "S101",
  # Don't require __init__.py
  "INP001",
  # Don't require trailing comma
  "COM812",
  # Don't require exception message in variable.
  "EM101",
  # Allow exception message.
  "TRY003",
  # Allow f-string in exception message.
  "EM102",
  # Don't require Author for TODOs.
  "TD002",
  # Don't require issue link for TODOs.
  "TD003",
  # Don't consider TODO an issue
  "FIX002",
  # Allow commented out code
  "ERA001",
]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "validator"]
